Simplify SDFFile
